requirements.txt
praw
instabot
python-resize-image